+++ eyes.appium.java/src/main/java/com/applitools/eyes/appium/AppiumFullPageCaptureAlgorithm.java
@@ -41,6 +41,7 @@ public class AppiumFullPageCaptureAlgorithm {
     private double pixelRatio;
     private BufferedImage stitchedImage;
     protected Location currentPosition;
+    protected WebElement scrollRootElement;
 
     // need to keep track of whether location and dimension coordinates returned by the driver
     // are already scaled to the pixel ratio, or are in "logical" pixels
@@ -89,13 +90,15 @@ public class AppiumFullPageCaptureAlgorithm {
                                           ImageProvider imageProvider, DebugScreenshotsProvider debugScreenshotsProvider,
                                           ScaleProviderFactory scaleProviderFactory, CutProvider cutProvider,
                                           EyesScreenshotFactory screenshotFactory, int waitBeforeScreenshots, WebElement cutElement,
-                                          Integer stitchingAdjustment) {
+                                          Integer stitchingAdjustment, WebElement scrollRootElement) {
 
         // ensure that all the scroll/position providers used by the superclass are the same object;
         // getting the current position for appium is very expensive!
         this(logger, testId, scrollProvider, scrollProvider, scrollProvider, imageProvider,
                 debugScreenshotsProvider, scaleProviderFactory, cutProvider, screenshotFactory,
                 waitBeforeScreenshots, cutElement, stitchingAdjustment);
+        this.scrollRootElement = scrollRootElement;
+        scrollProvider.setScrollRootElement(scrollRootElement);
     }
 
     protected RectangleSize captureAndStitchCurrentPart(Region partRegion) {
@@ -125,7 +128,9 @@ public class AppiumFullPageCaptureAlgorithm {
         // we modify the region by one pixel to make sure we don't accidentally get a pixel of the header above it
         Location newLoc = new Location(scrollViewRegion.getLeft(), scrollViewRegion.getTop() - scaleSafe(statusBarHeight) + 1);
         RectangleSize newSize = new RectangleSize(initialPartSize.getWidth(), scrollViewRegion.getHeight() - 1);
-        scrollViewRegion.setLocation(newLoc);
+        if (scrollRootElement == null) {
+            scrollViewRegion.setLocation(newLoc);
+        }
         scrollViewRegion.setSize(newSize);
 
         ((AppiumScrollPositionProvider) scrollProvider).setCutElement(cutElement);
@@ -151,7 +156,8 @@ public class AppiumFullPageCaptureAlgorithm {
             // here we make sure to say that the region we have scrolled to in the main screenshot
             // is also offset by 1, to match the change we made to the scrollViewRegion
             // We should set left = 0 because we need to a region from the start of viewport
-            currentPosition = new Location(currentPosition.getX(), currentPosition.getY() + 1 + stitchingAdjustment);
+            currentPosition = new Location(currentPosition.getX(), currentPosition.getY() + 1 + stitchingAdjustment
+                    + (scrollRootElement == null ? 0 : scaleSafe(statusBarHeight)));
 
             lastSuccessfulPartSize = captureAndStitchCurrentPart(regionToCrop);
         }
@@ -171,7 +177,8 @@ public class AppiumFullPageCaptureAlgorithm {
 
         cleanupStitch(originalStitchedState, currentPosition, lastSuccessfulPartSize, entireSize);
 
-        moveToTopLeft(xPos, endY + statusBarHeight + stitchingAdjustment, xPos, startY + statusBarHeight);
+        // 50 - extra offset to avoid overlay elements.
+        moveToTopLeft(xPos, endY + statusBarHeight + 50 + stitchingAdjustment, xPos, startY + statusBarHeight);
     }