+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [vNext]
 ### Added
 - Supporting OCR Feature. [Trello 2440](https://trello.com/c/FLGKnqIS)
+- Scroll root element functionality for iOS. [Trello 2347](https://trello.com/c/FDZ1oWF9)
 
 ## [3.192.0]
 ### Update